Changes to clean up the code, improve conceptual understanding
